Remove dtype cast

+++ tools/edm2_loss.py
@@ -75,7 +75,7 @@ class AdaptiveLossWeightMLP(nn.Module):
 
         if self.use_importance_weights:
             # min snr importance weights
-            all_timesteps = torch.arange(noise_scheduler.config.num_train_timesteps).to(device=device, dtype=dtype)
+            all_timesteps = torch.arange(noise_scheduler.config.num_train_timesteps).to(device=device)
             snr = torch.stack([noise_scheduler.all_snr[t] for t in all_timesteps])
 
             min_snr_gamma = (
